Fixes in Vortex modules playback and converting.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@839 e19a5227-2556-0410-816f-a35683127991

+++ src/core/plugins/players/ay/aym_parameters_helper.cpp
@@ -175,7 +175,7 @@ namespace
           throw MakeFormattedError(THIS_LINE, Module::ERROR_INVALID_PARAMETERS,
             Text::MODULE_ERROR_INVALID_LAYOUT, intParam);
         }
-        Chunk.Data[DataChunk::PARAM_LAYOUT] = intParam;
+        Chunk.Data[DataChunk::PARAM_LAYOUT] = static_cast<uint8_t>(intParam);
       }
     }
